+++ invocation.go
@@ -11,6 +11,7 @@ package di
 // Error will be returned as is.
 type Invocation interface{}
 
+// validateInvocation validates function.
 func validateInvocation(fn function) bool {
 	if fn.NumOut() == 0 {
 		return true
@@ -19,4 +20,17 @@ func validateInvocation(fn function) bool {
 		return true
 	}
 	return false
+}
+
+// parseInvocationParameters parses invocation and returns slice of nodes.
+func parseInvocationParameters(fn function, s schema) (params []*node, err error) {
+	for i := 0; i < fn.NumIn(); i++ {
+		in := fn.Type.In(i)
+		node, err := s.find(in, Tags{})
+		if err != nil {
+			return nil, err
+		}
+		params = append(params, node)
+	}
+	return params, nil
 }