+++ CHANGELOG.md
@@ -7,6 +7,16 @@ The format is based on
 project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html): TBD, use modules or another vendor system.
 
+## v1.7.0
+
+### Added
+
+- Added embed fields support.
+
+### Fixed
+
+- `di.Inject` now works with structs and pointers correctly.
+
 ## v1.6.3
 
 ### Fixed